+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @whatanime/design-tokens
 
+## 0.4.0
+
+### Minor Changes
+
+- add container component
+
 ## 0.3.0
 
 ### Minor Changes